# openapi.yaml
openapi: 3.0.3
info:
  title: Tienda de Videojuegos DEVCO
  description: |-
    API para una tienda de videojuegos, que permite realizar las operaciones
    básicas (CRUD). 
    
    Permite gestionar los videojuegos, las ordenes y los
    usuarios de la tienda.
  contact:
    email: asalazars@devco.com.co
  version: 1.0.0
...
paths:
...
  /usuario:
    get:
      tags:
        - usuario
      summary: Obtiene todos los usuarios de la tienda
      description: Obtiene todos los usuarios registrados de la tienda
      operationId: obtenerTodosUsuarios
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    nombre:
                      type: string
                      minLength: 1
                      example: usuario
                    email:
                      type: string
                      format: email
                      example: usuario@mail.com
                    rol:
                      type: string
                      example: cliente
    post:
      tags:
        - usuario
      summary: Crear un usuario
      description: Crear un usuario de tipo cliente
      operationId: crearUsuario
      requestBody:
        description: Usuario a crear
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  minLength: 1
                  example: usuario
                email:
                  type: string
                  format: email
                  example: usuario@mail.com
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: user123
                rol:
                  type: string
                  example: cliente
          application/xml:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  minLength: 1
                  example: usuario
                email:
                  type: string
                  format: email
                  example: usuario@mail.com
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: user123
                rol:
                  type: string
                  example: cliente
              xml:
                name: usuario
        required: true
      responses:
        '201':
          description: Creación exitosa
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  nombre:
                    type: string
                    minLength: 1
                    example: usuario
                  email:
                    type: string
                    format: email
                    example: usuario@mail.com
                  rol:
                    type: string
                    example: cliente
            application/xml:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  nombre:
                    type: string
                    minLength: 1
                    example: usuario
                  email:
                    type: string
                    format: email
                    example: usuario@mail.com
                  rol:
                    type: string
                    example: cliente
                xml:
                  name: usuario
        '400':
          description: Datos inválidos
...
  /usuario/login:
    post:
      tags:
        - usuario
      summary: Usuario se autentica en el sistema
      description: ''
      operationId: loginUsuario
      requestBody:
        description: Credenciales para acceder al sistema
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: admin@mail.com
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: admin123
          application/xml:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: admin@mail.com
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: admin123
              xml:
                name: usuario
        required: true
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: object
                properties:
                  nombre:
                    type: string
                    minLength: 1
                    example: admin
                  email:
                    type: string
                    format: email
                    example: admin@mail.com
                  rol:
                    type: string
                    example: admin
                  token:
                    type: string
                    example: token 
            application/xml:
              schema:
                type: object
                properties:
                  nombre:
                    type: string
                    minLength: 1
                    example: admin
                  email:
                    type: string
                    format: email
                    example: admin@mail.com
                  rol:
                    type: string
                    example: admin
                  token:
                    type: string
                    example: token
                xml:
                  name: usuario
        '400':
          description: Credenciales no suministradas
        '401':
          description: Credenciales inválidas
  /usuario/logout:
    post:
      tags:
        - usuario
      summary: Usuario cierra la sesión en el sistema
      description: ''
      operationId: logoutUsuario
      parameters: []
      responses:
        '200':
          description: Operación exitosa
      security:
        - bearerAuth: []
  /user/{idUsuario}:
    get:
      tags:
        - usuario
      summary: Obtiene un usuario por id
      description: Obtiene un único usuario
      operationId: obtenerUsuarioPorId
      parameters:
        - name: idUsuario
          in: path
          description: 'ID del usuario que se quiere obtener'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  nombre:
                    type: string
                    minLength: 1
                    example: usuario
                  email:
                    type: string
                    format: email
                    example: usuario@mail.com
                  rol:
                    type: string
                    example: cliente
            application/xml:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  nombre:
                    type: string
                    minLength: 1
                    example: usuario
                  email:
                    type: string
                    format: email
                    example: usuario@mail.com
                  rol:
                    type: string
                    example: cliente
                xml:
                  name: usuario
        '400':
          description: ID inválido
        '404':
          description: Usuario no encontrado
    put:
      tags:
        - usuario
      summary: Actualiza un usuario
      description: Actualiza un usuario registrado en la tienda, solo puede ser realizado por un usuario autenticado
      operationId: actualizarUsuario
      parameters:
        - name: idUsuario
          in: path
          description: 'ID del usuario que se quiere actualizar'
          required: true
          schema:
            type: integer
      requestBody:
        description: Usuario a actualizar
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  minLength: 1
                  example: admin
                email:
                  type: string
                  format: email
                  example: admin@mail.com
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: admin123
                rol:
                  type: string
                  example: admin
          application/xml:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  minLength: 1
                  example: admin
                email:
                  type: string
                  format: email
                  example: admin@mail.com
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: admin123
                rol:
                  type: string
                  example: admin
              xml:
                name: usuario
        required: true
      responses:
        '201':
          description: Actualización exitosa
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  nombre:
                    type: string
                    minLength: 1
                    example: admin
                  email:
                    type: string
                    format: email
                    example: admin@mail.com
                  rol:
                    type: string
                    example: admin
            application/xml:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  nombre:
                    type: string
                    minLength: 1
                    example: admin
                  email:
                    type: string
                    format: email
                    example: admin@mail.com
                  rol:
                    type: string
                    example: admin
                xml:
                  name: usuario
        '400':
          description: Datos inválidos
      security:
        - bearerAuth: []
    delete:
      tags:
        - usuario
      summary: Elimina un usuario
      description: Elimina un usuario registrado en la tienda, solo puede ser realizado por un usuario autenticado
      operationId: eliminarUsuario
      parameters:
        - name: idUsuario
          in: path
          description: 'ID del usuario que se quiere eliminar'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  nombre:
                    type: string
                    minLength: 1
                    example: admin
                  email:
                    type: string
                    format: email
                    example: admin@mail.com
                  rol:
                    type: string
                    example: admin
            application/xml:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  nombre:
                    type: string
                    minLength: 1
                    example: admin
                  email:
                    type: string
                    format: email
                    example: admin@mail.com
                  rol:
                    type: string
                    example: admin
                xml:
                  name: usuario
        '400':
          description: ID inválido
        '404':
          description: Usuario no encontrado
      security:
        - bearerAuth: []
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT